style: Update navbar styling and add horizontal rule in dropdown menu
Added some Morten-Hillbom magic to landing page

Refactor CSS in various components to improve navbar styling. Add a horizontal rule in the dropdown menu of the NavBarLoggedIn component for better visual separation. Also, tweak the padding and background color of the account dropdown menu items to enhance the user experience.