refactor: Extract InputQuill component from HowToMakeAssignments and AssignmentTextEditor
Prepared for making a save function (currently creates Quill.js component output that can later be loaded to display

Extract the InputQuill component from the HowToMakeAssignments and AssignmentTextEditor components to improve code modularity and reusability. This commit moves the InputQuill component code to its own file and updates the import statements in the HowToMakeAssignments and AssignmentTextEditor components to use the new file.