feat: Add CORS middleware to FastAPI application and add api file save in addition to local file save for assignment editor
The commit adds CORS middleware to the FastAPI application in order to allow cross-origin resource sharing. This is necessary to handle requests from different origins, such as when making API calls from a frontend application hosted on a different domain. The middleware is configured to allow all origins, credentials, methods, and headers. This improves the security and flexibility of the application.